namespaces and moves view statements into seperate module
